Some QOL updates
