+++ src/csv.ts
@@ -1,0 +1,96 @@
+
+// https://stackoverflow.com/a/1293163
+export function parse(strData: string, strDelimiter?: string | undefined): string[][] {
+    // Check to see if the delimiter is defined. If not,
+    // then default to comma.
+    strDelimiter = (strDelimiter || ",");
+
+    // Create a regular expression to parse the CSV values.
+    let objPattern = new RegExp(
+        (
+            // Delimiters.
+            "\\s*(\\" + strDelimiter + "|\\r?\\n|\\r|^)\\s*" +
+
+            // Quoted fields.
+            "(?:\"([^\"]*(?:\"\"[^\"]*)*)\"|" +
+
+            // Standard fields.
+            "([^\"\\" + strDelimiter + "\\r\\n]*))"
+        ),
+        "gi"
+    );
+
+
+    // Create an array to hold our data. Give the array
+    // a default empty first row.
+    let arrData: string[][] = [[]];
+
+    // Return empty arrays for empty string
+    if (!strData) return arrData;
+
+    // Create an array to hold our individual pattern
+    // matching groups.
+    let arrMatches = null;
+
+
+    // Keep looping over the regular expression matches
+    // until we can no longer find a match.
+    while (arrMatches = objPattern.exec(strData)) {
+
+        // Get the delimiter that was found.
+        var strMatchedDelimiter = arrMatches[1].replaceAll(/[^\r\n]/g, '');
+
+        // Check to see if the given delimiter has a length
+        // (is not the start of string) and if it matches
+        // field delimiter. If id does not, then we know
+        // that this delimiter is a row delimiter.
+        if (
+            strMatchedDelimiter.length && strMatchedDelimiter !== strDelimiter
+        ) {
+
+            // Since we have reached a new row of data,
+            // add an empty row to our data array.
+            arrData.push([]);
+        }
+
+        let strMatchedValue;
+
+        // Now that we have our delimiter out of the way,
+        // let's check to see which kind of value we
+        // captured (quoted or unquoted).
+        if (arrMatches[2]) {
+
+            // We found a quoted value. When we capture
+            // this value, unescape any double quotes.
+            strMatchedValue = arrMatches[2].replace(
+                new RegExp("\"\"", "g"),
+                "\""
+            );
+
+        } else {
+
+            // We found a non-quoted value.
+            strMatchedValue = arrMatches[3];
+
+        }
+
+
+        // Now that we have our value string, let's add
+        // it to the data array.
+        arrData[arrData.length - 1].push(strMatchedValue);
+    }
+
+    // Return the parsed data.
+    return arrData;
+}
+
+// https://stackoverflow.com/a/68146412
+export function stringify(arr: string[][]): string {
+    return arr.map(row =>
+        row
+            .map(String)  // convert every value to String
+            .map(v => v.replaceAll('"', '""'))  // escape double quotes
+            .map(v => `"${v}"`)  // quote it
+            .join(',')  // comma-separated
+    ).join('\r\n');
+}
