+++ src/renderer/index.ts
@@ -1,0 +1,1 @@
+export { HtmlView } from './html';
