+++ tsconfig.json
@@ -1,14 +1,16 @@
 {
 	"compilerOptions": {
-	  "module": "Node16",
-	  "target": "ES2022",
-	  "outDir": "out",
-	  "lib": [
-		"ES2022"
-	  ],
-	  "sourceMap": true,
-	  "rootDir": "src",
-	  "strict": true
-	}
-  }
-  
+		"module": "Node16",
+		"target": "ES2022",
+		"outDir": "out",
+		"lib": [
+			"ES2022"
+		],
+		"sourceMap": true,
+		"rootDir": "src",
+		"strict": true
+	},
+	"include": [
+		"src/**/*"
+	]
+}
